requirements.txt
gdown
torch
ultralytics
pyyaml
